+++ src/providers/ruby/requireCompletionProvider.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
 import { existsSync } from 'fs';
 import { basename, dirname, extname, posix, resolve, sep } from 'path';
-import { SORBET_FILE_DOCUMENT_SELECTOR } from '../../constants';
+import { PACKAGE_FILENAME, SORBET_FILE_DOCUMENT_SELECTOR } from '../../constants';
 import { ExtensionContext } from '../../extensionContext';
 
-export const TRIGGER_CHARACTERS: readonly string[] = ['"', "'", '/'];
+export const TRIGGER_CHARACTERS: readonly string[] = ['"', "'", posix.sep];
 
 export function registerRequireCompletionProvider({ disposables }: ExtensionContext): void {
   disposables.push(
@@ -48,13 +48,23 @@ export class RequireCompletionProvider implements vscode.CompletionItemProvider 
     const replaceRange = match.groups.path.length
       ? new vscode.Range(position.line, start, position.line, end)
       : undefined;
+
+    const dirCommitChars = [match.groups!.startQuote, posix.sep];
+    const fileCommitChars = [match.groups!.startQuote];
     return suggestions.map(([path, type]) => {
       const item = new vscode.CompletionItem(path, type);
       item.range = replaceRange;
-      item.commitCharacters = [match.groups!.startQuote];
+
       if (type === vscode.CompletionItemKind.Folder) {
-        item.commitCharacters.push('/');
+        item.commitCharacters = dirCommitChars;
+        item.insertText = path;
+        item.sortText = '0';
+      } else {
+        item.commitCharacters = fileCommitChars;
+        item.insertText = posix.join(hintPath.path, basename(path, '.rb'));
+        item.sortText = '1';
       }
+
       return item;
     });
 
@@ -81,24 +91,30 @@ export class RequireCompletionProvider implements vscode.CompletionItemProvider 
       hintPath: string,
     ): Promise<[string, vscode.CompletionItemKind][]> {
       const resolvedHintPath = resolve(baseDir, normalizeToRequire(hintPath));
-      const entries = await vscode.workspace.fs.readDirectory(vscode.Uri.file(resolvedHintPath));
-      if (!hintPath) {
-        entries.unshift(['..', vscode.FileType.Directory]);
+      let entries = await vscode.workspace.fs.readDirectory(vscode.Uri.file(resolvedHintPath));
+      if (baseDir === resolvedHintPath) {
+        const selfName = basename(document.fileName);
+        entries = entries.filter((entry) => entry[0] !== selfName);
       }
-      return entries
+      const paths = entries
         .map(([path, type]) => {
           let result: [string, vscode.CompletionItemKind] | undefined;
-          if (type === vscode.FileType.Directory) {
-            result = [posix.join(hintPath, path), vscode.CompletionItemKind.Folder];
-          } else if (extname(path) === '.rb') {
-            result = [posix.join(hintPath, basename(path, '.rb')), vscode.CompletionItemKind.File];
+          if (!path.startsWith('.')) {
+            if (type === vscode.FileType.Directory) {
+              result = [posix.join(hintPath, path), vscode.CompletionItemKind.Folder];
+            } else if (extname(path) === '.rb' && !path.endsWith(PACKAGE_FILENAME)) {
+              result = [posix.join(hintPath, path), vscode.CompletionItemKind.File];
+            }
           }
           return result;
         })
         .filter((e) => e !== undefined);
 
+      paths.unshift(['..', vscode.CompletionItemKind.Folder]);
+      return paths;
+
       function normalizeToRequire(path: string): string {
-        return sep !== '/' ? path.replaceAll(sep, '/') : path;
+        return sep !== posix.sep ? path.replaceAll(sep, posix.sep) : path;
       }
     }
   }