+++ src/providers/providers.ts
@@ -4,6 +4,7 @@ import { registerSorbetConfigHoverProvider } from './config/sorbetConfigHoverPro
 import { registerGemfileCodeLensProvider } from './gemfile/gemfileCodeLensProvider';
 import { registerGemfileCompletionProvider } from './gemfile/gemfileCompletionProvider';
 import { registerFoldingRangeProvider } from './ruby/foldingRangeProvider';
+import { registerRequireCompletionProvider } from './ruby/requireCompletionProvider';
 import { registerRequireRelativeCompletionProvider } from './ruby/requireRelativeCompletionProvider';
 import { registerRequireRelativeDefinitionProvider } from './ruby/requireRelativeDefinitionProvider';
 import { registerSelectionRangeProvider } from './ruby/selectionRangeProvider';
@@ -15,6 +16,7 @@ export function registerProviders(context: ExtensionContext): void {
   registerFoldingRangeProvider(context);
   registerGemfileCodeLensProvider(context);
   registerGemfileCompletionProvider(context);
+  registerRequireCompletionProvider(context);
   registerRequireRelativeCompletionProvider(context);
   registerRequireRelativeDefinitionProvider(context);
   registerSelectionRangeProvider(context);